address most of the review
